Update program_refresh.cpp
Your program has some minor bugs that are needed to be fixed asap. For example, when a polish user would like to use this application, it would be very unpleasant for him to see that polish diacritics are not here, so I am here to help you out with that kind of problem. Hope it helps.